refactor code and split it into separate files
